バリデーション機能、代替文書登録機能などの追加

+++ app/controllers/sections_controller.rb
@@ -1,11 +1,13 @@
 class SectionsController < ApplicationController
-  before_action :move_to_user_registration
-  before_action :set_owner_sections, only: [:index, :owner, :new, :edit, :show, :new_document]
+  before_action :move_to_user_registration, except: [:show]
+  before_action :set_owner_sections, except: [:destroy]
   
+  # 参加サイト一覧の表示
   def index
-    @participation_sections = current_user.participate_sections
+    @participation_sections = current_user.participate_sections.order("created_at DESC")
   end
 
+  # 管理サイト一覧の表示
   def owner
   end
 
@@ -15,16 +17,8 @@ class SectionsController < ApplicationController
   end
 
   def create
-    # binding.pry
-    # parent_section = Section.find(params[:section][:parent_id])
-    # @section = parent_section.children.new(section_name: params[:section][:section_name], user_id: current_user.id)
-    # if @section.save
-    #   redirect_to section_path(@section.id)
-    # else
-    #   render :new
-    # end
-
     @section = Section.new(section_params)
+    @section.user_id = current_user.id
     if @section.save
       redirect_to section_path(@section.id), notice: '分類を作成しました'
     else
@@ -32,8 +26,10 @@ class SectionsController < ApplicationController
     end
   end
 
+  # 編集権は作成者（管理者）のみ
   def edit
     @section = Section.find(params[:id])
+    redirect_to section_path(params[:id]) unless @section.user_id == current_user.id
   end
 
   def update
@@ -45,12 +41,18 @@ class SectionsController < ApplicationController
     end
   end
 
+  # 閲覧権（グループ参加者、グループ作成者、公開設定の場合は全員可）
   def show
     @section = Section.find(params[:id])
+    unless @section.disclosure_before_type_cast == 1 || @section.participate_users.where(id: current_user.id).present? || @section.user_id == current_user.id
+      return redirect_to root_path
+    end
   end
 
+  # 削除権は作成者（管理者）のみ
   def destroy
     @section = Section.find(params[:id])
+    return redirect_to section_path(params[:id]) unless @section.user_id == current_user.id
     if @section.destroy
       redirect_to root_path
     else
@@ -59,6 +61,7 @@ class SectionsController < ApplicationController
     end
   end
 
+  # 区分一覧画面から新規document作成
   def new_document
     @document = Document.new(section_id: params[:id])
     render "documents/new"
@@ -66,12 +69,16 @@ class SectionsController < ApplicationController
 
   private
   def section_params
-    params.require(:section).permit(:section_name, :disclosure, :gist, participate_user_ids: []).merge(user_id: current_user.id)
+    params.require(:section).permit(:section_name, :disclosure, :gist, participate_user_ids: [])
   end
   
   def move_to_user_registration
     return redirect_to new_user_registration_path unless user_signed_in?
   end
+
+  # def move_to_root_path(section)
+  #   if section.where(section.id: current_user.participate_sections.ids)
+  # end
 
   def set_owner_sections
     return @owner_sections = [] unless user_signed_in?
